+++ geotrek/infrastructure/management/commands/loadsignage.py
@@ -1,4 +1,4 @@
-# -*- coding: utf8 -*-
+# -*- coding: utf-8 -*-
 
 import os.path
 from optparse import make_option