+++ geotrek/trekking/management/commands/sync_rando.py
@@ -47,6 +47,14 @@ class ZipTilesBuilder(object):
         self.close_zip = close_zip
         self.zipfile = ZipFile(filepath, 'w')
         self.tm = TilesManager(**builder_args)
+
+        if not isinstance(settings.MOBILE_TILES_URL, str) and len(settings.MOBILE_TILES_URL) > 1:
+            for url in settings.MOBILE_TILES_URL[1:]:
+                args = builder_args
+                args['tiles_url'] = url
+                args['tile_format'] = self.format_from_url(args['tiles_url'])
+                self.tm.add_layer(TilesManager(**args), opacity=1)
+
         self.tiles = set()
 
     def format_from_url(self, url):
@@ -65,7 +73,7 @@ class ZipTilesBuilder(object):
 
     def run(self):
         for tile in self.tiles:
-            name = '{0}/{1}/{2}.png'.format(*tile)
+            name = '{0}/{1}/{2}{ext}'.format(*tile, ext=settings.MOBILE_TILES_EXTENSION or self.tm._tile_extension)
             try:
                 data = self.tm.tile(tile)
             except DownloadError:
@@ -109,7 +117,7 @@ class Command(BaseCommand):
         """
         zipname = os.path.join('zip', 'tiles', 'global.zip')
 
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m**\x1b[0m \x1b[1m{name}\x1b[0m ...".format(name=zipname), ending="")
             self.stdout.flush()
 
@@ -134,7 +142,7 @@ class Command(BaseCommand):
         """
         zipname = os.path.join('zip', 'tiles', '{pk}.zip'.format(pk=trek.pk))
 
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m**\x1b[0m \x1b[1m{name}\x1b[0m ...".format(name=zipname), ending="")
             self.stdout.flush()
 
@@ -165,7 +173,7 @@ class Command(BaseCommand):
         tiles.run()
 
     def sync_view(self, lang, view, name, url='/', params={}, zipfile=None, **kwargs):
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m{lang}\x1b[0m \x1b[1m{name}\x1b[0m ...".format(lang=lang, name=name), ending="")
             self.stdout.flush()
         fullname = os.path.join(self.tmp_root, name)
@@ -175,16 +183,16 @@ class Command(BaseCommand):
         request.user = AnonymousUser()
         try:
             response = view(request, **kwargs)
+            if hasattr(response, 'render'):
+                response.render()
         except Exception as e:
             self.successfull = False
-            if self.verbosity == '2':
+            if self.verbosity == 2:
                 self.stdout.write(u"\x1b[3D\x1b[31mfailed ({})\x1b[0m".format(e))
             return
-        if hasattr(response, 'render'):
-            response.render()
         if response.status_code != 200:
             self.successfull = False
-            if self.verbosity == '2':
+            if self.verbosity == 2:
                 self.stdout.write(u"\x1b[3D\x1b[31;1mfailed (HTTP {code})\x1b[0m".format(code=response.status_code))
             return
         f = open(fullname, 'w')
@@ -196,7 +204,7 @@ class Command(BaseCommand):
         f.close()
         if zipfile:
             zipfile.write(fullname, name)
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[3D\x1b[32mgenerated\x1b[0m")
 
     def sync_json(self, lang, viewset, name, zipfile=None, params={}, as_view_args=[], **kwargs):
@@ -291,7 +299,7 @@ class Command(BaseCommand):
         shutil.copyfile(src, dst)
         if zipfile:
             zipfile.write(dst, os.path.join(url, name))
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m{lang}\x1b[0m \x1b[1m{url}/{name}\x1b[0m \x1b[32mcopied\x1b[0m".format(lang=lang, url=url, name=name))
 
     def sync_static_file(self, lang, name):
@@ -346,7 +354,7 @@ class Command(BaseCommand):
         if self.categories:
             self.sync_trek_touristiccontents(lang, trek, zipfile=self.zipfile)
 
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m{lang}\x1b[0m \x1b[1m{name}\x1b[0m ...".format(lang=lang, name=zipname),
                               ending="")
 
@@ -370,7 +378,7 @@ class Command(BaseCommand):
             stat = os.stat(oldzipfilename)
             os.utime(zipfilename, (stat.st_atime, stat.st_mtime))
 
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             if uptodate:
                 self.stdout.write(u"\x1b[3D\x1b[32munchanged\x1b[0m")
             else:
@@ -420,7 +428,7 @@ class Command(BaseCommand):
 
         self.sync_tourism(lang)
 
-        if self.verbosity == '2':
+        if self.verbosity == 2:
             self.stdout.write(u"\x1b[36m{lang}\x1b[0m \x1b[1m{name}\x1b[0m ...".format(lang=lang, name=zipname), ending="")
 
         self.close_zip(self.zipfile, zipname)
@@ -642,7 +650,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         self.successfull = True
-        self.verbosity = options.get('verbosity', '1')
+        self.verbosity = options.get('verbosity', 1)
         if len(args) < 1:
             raise CommandError(u"Missing parameter destination directory")
         self.dst_root = args[0].rstrip('/')
@@ -678,8 +686,12 @@ class Command(BaseCommand):
         else:
             self.portal = []
 
+        if isinstance(settings.MOBILE_TILES_URL, str):
+            tiles_url = settings.MOBILE_TILES_URL
+        else:
+            tiles_url = settings.MOBILE_TILES_URL[0]
         self.builder_args = {
-            'tiles_url': settings.MOBILE_TILES_URL,
+            'tiles_url': tiles_url,
             'tiles_headers': {"Referer": self.referer},
             'ignore_errors': True,
             'tiles_dir': os.path.join(settings.DEPLOY_ROOT, 'var', 'tiles'),
@@ -702,7 +714,7 @@ class Command(BaseCommand):
 
         self.rename_root()
 
-        if self.verbosity >= '1':
+        if self.verbosity >= 1:
             self.stdout.write('Done')
 
         if not self.successfull: